moved files and css around to clean up code, added navigation to createRequest instead of href
