+++ src/components/Interface/Style.css
@@ -20,7 +20,6 @@ div {
     font-family: Arial;
 }
 
-
 .slide-post {
     padding: 20px;
 }
@@ -34,7 +33,6 @@ div {
     height: 50px;
     width: 100px;
     background: linear-gradient(-90deg, #962D38 50%, #EAB547 50%);
-
 }
 
 .purple-black-button {
@@ -68,4 +66,31 @@ div {
 .speaker-logo {
     max-width: 35px;
     max-height: 35px;
+}
+
+.formGroup {
+    margin-bottom: 15px;
+}
+
+.requestTableDiv {
+    height: 55vh;
+    overflow-y: auto;
+    overflow-x: hidden;
+}
+
+.requestTable {
+    border: 4px solid #ddd;
+    border-collapse: collapse;
+    border-radius: 4px;
+    border-spacing: 0;
+}
+
+.requestTableRow {
+    vertical-align: middle;
+}
+
+.requestDetailButton {
+    text-align: center;
+    margin: auto;
+    display: block;
 }
