--- src/components/Request/RequestComp.css
@@ -1,3 +1,0 @@
-.formGroup {
-    margin-bottom: 15px;
-}
