+++ src/components/Request/RequestComp.js
@@ -6,12 +6,16 @@ import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 import Button from "react-bootstrap/Button";
 import "bootstrap/dist/css/bootstrap.min.css";
-
 import RequestTable from "./RequestTable";
-import classes from "./RequestComp.css";
+import {useNavigate} from "react-router-dom";
 import "../Interface/Style.css"
 
 export default function RequestComp() {
+    const navigate = useNavigate();
+    const handleCreateRequest = function () {
+        navigate('/create');
+    }
+
     return (
         <Container className="priority-announcement">
             <Card className="shadow-sm">
@@ -21,7 +25,7 @@ export default function RequestComp() {
                             <h2>Your Requests</h2>
                         </div>
                     </Row>
-                    <Row className={classes.formGroup}>
+                    <Row className="formGroup">
                         <Tabs defaultActiveKey="pending" id="table-tabs" className="mb-3">
                             <Tab eventKey="pending" title="Pending">
                                 <RequestTable status="Pending"/>
@@ -36,7 +40,7 @@ export default function RequestComp() {
                     </Row>
                     <Row className="float-end">
                         <Col>
-                            <Button href="#">Create New Request</Button>
+                            <Button onClick={handleCreateRequest}>Create New Request</Button>
                         </Col>
                     </Row>
                 </Card.Body>
