+++ src/components/Request/RequestCreate.js
@@ -4,7 +4,6 @@ import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
-
 import "../Interface/Style.css";
 
 export default function RequestCreate() {
@@ -58,7 +57,7 @@ export default function RequestCreate() {
                             <Row className="float-end">
                                 <Col>
                                     <Button variant="primary" type="submit">
-                                        create
+                                        Submit
                                     </Button>
                                 </Col>
                             </Row>
