--- src/components/Request/RequestTable.css
@@ -1,20 +1,0 @@
-.row {
-    vertical-align: middle;
-}
-
-.button {
-    text-align: center;
-}
-
-.reactTable {
-    height: 55vh;
-    overflow-y: auto;
-    overflow-x: hidden;
-}
-
-.bdr {
-    border: 4px solid #ddd;
-    border-collapse: separate;
-    border-radius: 4px;
-    border-spacing: 0px;
-}
