+++ src/pages/RequestCreate.js
@@ -1,5 +1,5 @@
 import {Fragment, React} from "react";
-import RequestCreateComp from "../components/RequestCreate/RequestCreate"
+import RequestCreateComp from "../components/Request/RequestCreate"
 import "bootstrap/dist/css/bootstrap.min.css";
 
 export default function RequestCreate() {
